requirements.txt
asgiref==3.9.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.4
django-environ==0.12.0
django-silk==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.3.0
gprof2dot==2025.4.14
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.9.0
requests==2.32.4
sqlparse==0.5.3
stripe==12.3.0
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0